Use a single nidas tool to consolidate SConscript boilerplate, such as registering nidas apps, adding library paths, retrieving nidas app targets, and building programs with install and clean aliases.  The nidas tool is now loaded in SConscript files which need access to those wrapper methods.
Remove the duplication of eol_scons tool variables from the top-level
SConstruct.  Instead share nidas.conf between both top-level and eol_scons
Variables() instances, and use both to generate the help text.  I tried
combining the nidas opts with the global opts, but the PREFIX and MACHS
settings are modified when passed to lower SConscript calls, in which case
they cannot be re-updated from the global opts.

nidas/apps/SConscript now uses duplicate=0 for the config and auto_cal
builds.

nidas/apps/config/site_scons is not needed so remove svn:externals.

tests/UTime/SConscript missed dependency on utime program.

Remove references to netcdfrpc outputs in the test xml files, so the tests can
run and pass even when netcdf server is not built.

Replace nidas/apps/config/SConstruct with a redirect to the SConscript
file, so the scripting for building configedit is only maintained in one
place.  The SConscript file builds configedit both inside and outside the
nidas source tree, thanks to the nidas tool.